renamed component, without *page suffix. Removed AcademicIdentitiesPage
